Make splitFastqBarcodes gzip-transparent
Reuse the NucProcess.open_file_r function.
